Pin Rust/Java versions, cache Gradle dependencies in CI
Make sure the Rust and Java versions stay stable so that folks
running pipelines get the same results.

Pre-run the Gradle command so the `~/.konan` (Kotlin Native),
`~/.gradle`, `/usr/local/cargo/`, and `/usr/local/rustup/` caches are
all warm.

I updated the pipelines to (1) only build the Docker image if a new one
is needed but also (2) make a way to manually build a new image.

Closes https://gitlab.com/trixnity/uniffi-kotlin-multiplatform-bindings/-/issues/19