fix: Build bindings from the crate directory
Right now, the bindgen task always runs from the root of the project.
Unfortunately that means that if the Cargo.toml file is in a `/uniffi`
folder, you need to also have a Cargo.toml workspace file at the root of
the project to point to the nested `uniffi` folder.

This change follows the convention from the other commands which run
from the crate directory (and it doesn't change the API).